Issue-127: be more explicit about creating indices
